Use custom `UtilLinuxLogger` instead `syslog`
`syslog` used to be part of StdLib. But beginning with  Ruby 3.4, it was
extracted into default gems. This is problematic when Ruby is executed
with `--disable-gems`, because there is no easy way to require `syslog`.

Therefore, this commit introduces custom `UtilLinuxLogger`. This use
`logger` command from util-linux and implemnts just minimal `Syslog`
interface required.

While this could have been used just as a fallback, when `syslog` can't
be required, it will be easier to use it every time.

Fixes #12